% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{choose_k}
\alias{choose_k}
\title{Choose number of outliers k}
\usage{
choose_k(
  x,
  y,
  k_vec,
  l = 0,
  intercept = TRUE,
  ic_ind = 3,
  b_init = NULL,
  tol_ls = 1e-06,
  MaxIter_ls = 100,
  bd_const = 2,
  time_limit = 600,
  suppress_msg = TRUE,
  tol_iht = 1e-06,
  MaxIter_iht = 2000,
  int_tol = 1e-08,
  MIPGap = 0.001,
  MIPFocus = 3
)
}
\arguments{
\item{k_vec}{candidate k values}

\item{l}{if l = 0, use IHT without local search.}

\item{ic_ind}{which IC to be used.}

\item{int_tol}{tolerance for integer in gurobi}

\item{MIPGap}{termination condition for gurobi}

\item{MIPFocus}{strategies for gurobi}
}
\value{
selected k (scalar or vector, len = length(ic_ind))
}
\description{
Choose number of outliers k
}
