% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{choose_lambda}
\alias{choose_lambda}
\title{Choose tuning parameters for L1 method}
\usage{
choose_lambda(
  x,
  y,
  intercept = TRUE,
  lambda_seq = NULL,
  ic_ind = 3,
  parallel = FALSE,
  nlambda = 100,
  lambda_min_ratio = 1e-04
)
}
\arguments{
\item{lambda_seq}{candidate k values}

\item{ic_ind}{which IC to be used.}

\item{lambda_min_ratio}{}
}
\description{
Choose tuning parameters for L1 method
}
