% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_lambda_seq}
\alias{get_lambda_seq}
\title{The function generate candidate tuning parameters for L1 method}
\usage{
get_lambda_seq(x, y, intercept, nlambda = 100, lambda_min_ratio = 1e-04)
}
\description{
The function generate candidate tuning parameters for L1 method
}
