% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{huber_reg}
\alias{huber_reg}
\title{Huber regression based on Yi and Huang (2016)}
\usage{
huber_reg(x, y, intercept = TRUE, huber_thres = 1.345)
}
\arguments{
\item{huber_thres}{}
}
\value{
estimated coefficient
}
\description{
Yi, C. and Huang, J. (2016) Semismooth Newton Coordinate Descent Algorithm
for Elastic-Net Penalized Huber Loss Regression and Quantile Regression
}
