% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{huber_reg_cvxr}
\alias{huber_reg_cvxr}
\title{Huber regression}
\usage{
huber_reg_cvxr(x, y, intercept = TRUE, huber_thres = 1.345)
}
\arguments{
\item{huber_thres}{}
}
\value{
estimated coefficient
}
\description{
Huber regression
}
