% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{iht}
\alias{iht}
\title{Iterative Hard-thresholding (IHT) (Algorithm 1)}
\usage{
iht(
  x,
  y,
  k,
  intercept = TRUE,
  b0 = NULL,
  b_init = NULL,
  tol = 1e-06,
  MaxIter = 2000
)
}
\arguments{
\item{b0}{Null hypothesis}

\item{tol}{Convergence criteria}

\item{MaxIter}{}
}
\value{
A list contains estimation results
\item{b} Estimated coefficient
\item{a} Estimated fixed effects
\item{d} = 0 if detected outlier; = 1 if not.
\item{t}
}
\description{
Iterative Hard-thresholding (IHT) (Algorithm 1)
}
