% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{local_search}
\alias{local_search}
\title{Local Combinatorial Search (Algorithm 2)}
\usage{
local_search(
  x,
  y,
  k,
  l,
  intercept = TRUE,
  b_init = NULL,
  tol = 1e-06,
  MaxIter = 100,
  bd_const = 2,
  time_limit = 300,
  suppress_msg = TRUE,
  tol_iht = 1e-07,
  MaxIter_iht = 2000,
  int_tol = 1e-08,
  MIPGap = 0.001,
  MIPFocus = 3
)
}
\arguments{
\item{l}{local exactness level}

\item{int_tol}{tolerance for integer in gurobi}

\item{MIPGap}{termination condition for gurobi}

\item{MIPFocus}{strategies for gurobi}
}
\description{
Local Combinatorial Search (Algorithm 2)
}
