% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{local_search_mio}
\alias{local_search_mio}
\title{local combinatorial search MIO problem}
\usage{
local_search_mio(
  x,
  y,
  k,
  l,
  intercept = TRUE,
  start_value,
  bd_const = 2,
  time_limit = 300,
  suppress_msg = TRUE,
  int_tol = 1e-08,
  MIPGap = 0.001,
  MIPFocus = 3
)
}
\arguments{
\item{start_value}{(b, a, gamma) solution to be checked}

\item{bd_const}{parameter for bounds of variables for Gurobi solver}

\item{int_tol}{tolerance for integer in gurobi}

\item{MIPGap}{termination condition for gurobi}

\item{MIPFocus}{strategies for gurobi}
}
\description{
local combinatorial search MIO problem
}
