% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lsfit_robust_mio}
\alias{lsfit_robust_mio}
\title{Robust Linear Regression with L0 regularization via MIO}
\usage{
lsfit_robust_mio(
  x,
  y,
  k,
  intercept = TRUE,
  b0 = NULL,
  bd_const = 2,
  time_limit = 300,
  start_value = NULL,
  suppress_msg = TRUE,
  tol_iht = 1e-06,
  MaxIter_iht = 2000,
  int_tol = 1e-08,
  MIPGap = 0.001,
  MIPFocus = 3
)
}
\arguments{
\item{b0}{Null hypothesis}

\item{bd_const}{parameter for bounds of variables for Gurobi solver}

\item{time_limit}{time limit for Gurobi solver}

\item{start_value}{start value for Gurobi solver; by default using IHT}

\item{int_tol}{tolerance for integer in gurobi}

\item{MIPGap}{termination condition for gurobi}

\item{MIPFocus}{strategies for gurobi}
}
\value{
A list contains estimation results
\item{b} Estimated coefficient
\item{a} Estimated fixed effects
\item{d} = 0 if detected outlier; = 1 if not.
\item{t}
\item{sol_status}
}
\description{
Robust Linear Regression with L0 regularization via MIO
}
